// apps/web/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        StudentDetailDto: {
            id: string;
            name: string;
            yearLevel: string;
            canEdit: boolean;
            version: number;
            firstEnrolledOn: string | null;
            guardianName?: string;
            guardianRelationship?: string;
            guardianPhone?: string;
            guardianEmail?: string;
            guardianWechat?: string;
            preferredChannel?: string;
            preferredLanguage?: string;
            learningGoals?: string;
            preferredTimes?: string;
            interestedSubjects?: string;
            teachingRecords: components["schemas"]["TeachingRecordDto"][];
        };
        UpdateStudentDto: {
            guardianName?: string;
            guardianRelationship?: string;
            guardianPhone?: string;
            guardianEmail?: string;
            guardianWechat?: string;
            /** @enum {string} */
            preferredChannel?: "" | "EMAIL" | "PHONE" | "SMS" | "WECHAT";
            /** @enum {string} */
            preferredLanguage?: "en-AU" | "zh-CN";
            learningGoals?: string;
            preferredTimes?: string;
            interestedSubjects?: string;
            /** @example 2026-09-17 */
            firstEnrolledOn?: string;
            name?: string;
            /** @enum {string} */
            yearLevel?: "Foundation" | "Year 1" | "Year 2" | "Year 3" | "Year 4" | "Year 5" | "Year 6" | "Year 7" | "Year 8" | "Year 9" | "Year 10" | "Year 11" | "Year 12" | "Not assessed";
            expectedVersion: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}